added skeletons for transaction list

+++ frontend/src/lib/index.ts
@@ -1,49 +1,49 @@
 import { PUBLIC_API_URL } from "$env/static/public";
 
 export function getApiUrl(endpoint: string): string {
-	return `${PUBLIC_API_URL}/${endpoint}`;
+    return `${PUBLIC_API_URL}/${endpoint}`;
 }
 
 export function getFormattedDate(date: Date): string {
-	return new Date(date).toLocaleDateString("default", {
-		weekday: "short",
-		year: "numeric",
-		month: "long",
-		day: "numeric",
-	});
+    return new Date(date).toLocaleDateString("default", {
+        weekday: "short",
+        year: "numeric",
+        month: "long",
+        day: "numeric",
+    });
 }
 
 export function getFormattedDateShort(date: Date): string {
-	return new Date(date).toLocaleDateString("default", {
-		month: "long",
-		day: "numeric",
-	});
+    return new Date(date).toLocaleDateString("default", {
+        month: "short",
+        day: "numeric",
+    });
 }
 
 export async function authorizeFetch(
-	endpoint: string,
-	accessToken: string,
-	method: string = "GET",
+    endpoint: string,
+    accessToken: string,
+    method: string = "GET",
 ) {
-	return await fetch(getApiUrl(endpoint), {
-		method: method,
-		headers: {
-			Authorization: `Bearer ${accessToken}`,
-		},
-	});
+    return await fetch(getApiUrl(endpoint), {
+        method: method,
+        headers: {
+            Authorization: `Bearer ${accessToken}`,
+        },
+    });
 }
 
 export async function authorizePost(
-	endpoint: string,
-	accessToken: string,
-	data: string,
+    endpoint: string,
+    accessToken: string,
+    data: string,
 ) {
-	return await fetch(getApiUrl(endpoint), {
-		method: "POST",
-		headers: {
-			Authorization: `Bearer ${accessToken}`,
-			"Content-Type": "application/json",
-		},
-		body: data,
-	});
+    return await fetch(getApiUrl(endpoint), {
+        method: "POST",
+        headers: {
+            Authorization: `Bearer ${accessToken}`,
+            "Content-Type": "application/json",
+        },
+        body: data,
+    });
 }
